# BACKEND/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "cuidarteplus API (español)",
    "version": "1.0.0"
  },
  "servers": [{ "url": "http://localhost:4444" }],
  "paths": {
...
    "/examenes": {
      "get": {
        "summary": "Listar exámenes médicos",
        "description": "Lista todos los exámenes médicos. Médicos y admins ven todos los exámenes. Pacientes solo ven sus propios exámenes.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "No autorizado" },
          "403": { "description": "No tiene permisos" }
        }
      },
      "post": {
        "summary": "Crear examen médico",
        "description": "Crea un nuevo examen médico. Requiere rol de médico o admin.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ExamenMedicoCrear" }
            }
          }
        },
        "responses": {
          "201": { "description": "Examen creado" },
          "401": { "description": "No autorizado" },
          "403": { "description": "No tiene permisos" }
        }
      }
    },
    "/examenes/{id}": {
      "get": {
        "summary": "Obtener examen médico",
        "description": "Obtiene la información de un examen médico específico. Médicos y admins pueden ver cualquier examen. Pacientes solo pueden ver sus propios exámenes.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "No autorizado" },
          "403": { "description": "No tiene permisos" },
          "404": { "description": "Examen no encontrado" }
        }
      },
      "put": {
        "summary": "Actualizar examen médico",
        "description": "Actualiza un examen médico. Requiere rol de médico o admin.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamenMedicoActualizar"
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Error de validación" },
          "401": { "description": "No autorizado" },
          "403": { "description": "No tiene permisos" },
          "404": { "description": "Examen no encontrado" }
        }
      },
      "delete": {
        "summary": "Eliminar examen médico",
        "description": "Elimina un examen médico. Solo admin puede eliminar exámenes.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Examen eliminado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deleted": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "401": { "description": "No autorizado" },
          "403": { "description": "No tiene permisos" },
          "404": { "description": "Examen no encontrado" }
        }
      }
    },
    "/examenes/paciente/{paciente_id}": {
      "get": {
        "summary": "Obtener exámenes por paciente",
        "description": "Obtiene todos los exámenes médicos de un paciente específico. Médicos y admins pueden ver exámenes de cualquier paciente. Pacientes solo pueden ver sus propios exámenes.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "paciente_id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "No autorizado" },
          "403": { "description": "No tiene permisos" }
        }
      }
    },
...
    "/documentos/examen/{examen_id}": {
      "get": {
        "summary": "Listar documentos de un examen",
        "description": "Lista todos los documentos asociados a un examen médico. Médicos y admins pueden ver documentos de cualquier examen. Pacientes solo pueden ver documentos de sus propios exámenes.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "examen_id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "No autorizado" },
          "403": { "description": "No tiene permisos" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExamenMedicoCrear": {
        "type": "object",
        "properties": {
          "tipo_examen_medico_id": { "type": "integer" },
          "paciente_id": { "type": "integer" },
          "diagnosis": { "type": "string" },
          "tratamiento": { "type": "string" },
          "observaciones": { "type": "string" },
          "notas": { "type": "string" }
        },
        "required": ["tipo_examen_medico_id", "paciente_id"]
      },
      "ExamenMedicoActualizar": {
        "type": "object",
        "properties": {
          "tipo_examen_medico_id": { "type": "integer" },
          "paciente_id": { "type": "integer" },
          "diagnosis": { "type": "string" },
          "tratamiento": { "type": "string" },
          "observaciones": { "type": "string" },
          "notas": { "type": "string" }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenido del endpoint /autenticacion/login. Incluye: usuarioId, nombreUsuario, rolId, rolNombre, pacienteId, roles"
      }
    }
  }
}